# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Chatbot Law API",
    "version": "1.0.0",
    "description": "API documentation for the AI Chatbot for Education Laws."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/chat-room/{chat_id}/messages": {
      "get": {
        "summary": "Lấy lịch sử chat của phòng",
        "description": "Lấy tất cả tin nhắn (question-answer pairs) trong một phòng chat.",
        "tags": ["ChatRoom"],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "description": "ID của phòng chat",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lịch sử chat của phòng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChatMessage" }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token vào đây. Ví dụ: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    },
    "schemas": {
...
      "ChatMessage": {
        "type": "object",
        "properties": {
          "question": { "type": "string" },
          "answer": { "type": "string" },
          "question_date": { "type": "string", "format": "date-time" },
          "answer_date": { "type": "string", "format": "date-time" },
          "chat_id": { "type": "string" }
        }
      },
...
    }
  }
}